backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.4
django-filter==22.1
django-phonenumber-field==7.0.1
django-twilio==0.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-dotenv
django-celery-results
idna==3.4
importlib-metadata==5.1.0
kombu==5.2.4
Markdown==3.4.1
phonenumbers==8.13.3
prompt-toolkit==3.0.36
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
twilio==7.16.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zipp==3.11.0